Add ProgressWrapper in a separate component
