+++ src/components/Progress.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import style from './../styles.css'
+import ProgressWrapper from './ProgressWrapper'
 
 export default class Progress extends React.PureComponent {
   componentDidMount() {
@@ -24,17 +25,6 @@ export default class Progress extends React.PureComponent {
   }
 }
 
-const ProgressWrapper = (props) => (
-  <div className={style.autoHide} style={{ ...styles.progress, ...getProgressWrapperStyle(props) }}>
-    {props.children}
-  </div>
-)
-
-const getProgressWrapperStyle = ({ width, pause, bufferAction }) => ({
-  width: `${width * 100}%`,
-  opacity: pause && !bufferAction ? 0 : 1
-})
-
 const animationPlayState = (pause) => pause ? 'paused' : 'running'
 
 const getProgressStyle = ({ active, defaultInterval, pause }) => {
@@ -47,15 +37,6 @@ const getProgressStyle = ({ active, defaultInterval, pause }) => {
       return { width: 0 }
     default:
       return { width: 0 }
-  }
-}
-
-const styles = {
-  progress: {
-    height: 2,
-    maxWidth: '100%',
-    background: '#555',
-    margin: 2
   }
 }
